Some more plausibilization in parsing
When we want to parse a steam, that failed when that stream metadata
contained the word "stream" itself, because we're using really simple
string parsing instead of relying on the XRef directory, which would
actually be the proper way of doing things. We kind of "fix" it by also
requiring the "endstream" keyword to be in there as well, but that's not
really a fix. Encountered this purely by accident when trying to parse a
Chromium-printed PDF document that contained a URI which contained
"stream".